Unreviewed build fix.
Source/WebCore:

* platform/graphics/chromium/cc/CCLayerTreeHostImpl.cpp:
(WebCore::CCLayerTreeHostImpl::removePassesWithCachedTextures):

Source/WebKit/chromium:

* tests/CCLayerTreeHostImplTest.cpp:

Canonical link: https://commits.webkit.org/107176@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@120564 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/chromium/cc/CCLayerTreeHostImpl.cpp
@@ -418,7 +418,7 @@ void CCLayerTreeHostImpl::removeRenderPassesRecursive(CCRenderPassList& passes, 
 
 void CCLayerTreeHostImpl::removePassesWithCachedTextures(CCRenderPassList& passes, CCRenderPassList& skippedPasses)
 {
-    for (ssize_t passIndex = passes.size() - 1; passIndex >= 0; --passIndex) {
+    for (int passIndex = passes.size() - 1; passIndex >= 0; --passIndex) {
         CCRenderPass* currentPass = passes[passIndex].get();
         const CCQuadList& quadList = currentPass->quadList();
         CCQuadList::constBackToFrontIterator quadListIterator = quadList.backToFrontBegin();
